requirements.txt
discord.py==1.7.3
discord-py-interactions==5.5.0